feat: change js files to jsx
