+++ frontend/src/components/auth/ProtectedRoute.jsx
