+++ frontend/src/index.jsx
@@ -1,0 +1,32 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import { Provider } from "react-redux";
+import { BrowserRouter } from "react-router-dom";
+
+import { ModalProvider, Modal } from "./context/Modal";
+import configureStore from "./store";
+import * as sessionActions from "./store/session";
+import App from "./App";
+
+import "./index.css";
+
+const store = configureStore();
+
+if (process.env.NODE_ENV !== "production") {
+  window.store = store;
+  window.sessionActions = sessionActions;
+}
+
+ReactDOM.render(
+  <React.StrictMode>
+    <ModalProvider>
+      <Provider store={store}>
+        <BrowserRouter>
+          <App />
+          <Modal />
+        </BrowserRouter>
+      </Provider>
+    </ModalProvider>
+  </React.StrictMode>,
+  document.getElementById("root")
+);
